refactor: 도메인 패키지 분리
- 예약, 예약 시간, 테마 관련 클래스를 도메인 패키지 하위에 분리한다
