+++ src/main/java/roomescape/controller/dto/response/RoomThemeResponse.java
@@ -1,6 +1,6 @@
 package roomescape.controller.dto.response;
 
-import roomescape.domain.RoomTheme;
+import roomescape.domain.roomtheme.RoomTheme;
 
 public record RoomThemeResponse(long id, String name, String description, String thumbnail) {
