+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
-import roomescape.domain.AvailableReservationTime;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservationtime.AvailableReservationTime;
+import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.exception.custom.ExistedDuplicateValueException;
 import roomescape.exception.custom.InUseException;
 import roomescape.exception.custom.NotExistedValueException;
