+++ src/main/java/roomescape/service/RoomThemeService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.RoomThemeDao;
-import roomescape.domain.PopularThemeSelectionCriteria;
-import roomescape.domain.RoomTheme;
+import roomescape.domain.roomtheme.PopularThemeSelectionCriteria;
+import roomescape.domain.roomtheme.RoomTheme;
 import roomescape.exception.custom.ExistedDuplicateValueException;
 import roomescape.exception.custom.InUseException;
 import roomescape.exception.custom.NotExistedValueException;
