+++ src/test/java/roomescape/domain/AvailableReservationTimeTest.java
@@ -7,6 +7,8 @@ import java.time.LocalTime;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.reservationtime.AvailableReservationTime;
+import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.exception.custom.InvalidInputException;
 
 class AvailableReservationTimeTest {
