+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -7,6 +7,9 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservationtime.ReservationTime;
+import roomescape.domain.roomtheme.RoomTheme;
 import roomescape.exception.custom.InvalidInputException;
 
 class ReservationTest {
