+++ src/test/java/roomescape/domain/RoomThemeTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.roomtheme.RoomTheme;
 import roomescape.exception.custom.InvalidInputException;
 
 class RoomThemeTest {
