+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -13,8 +13,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.domain.AvailableReservationTime;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservationtime.AvailableReservationTime;
+import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.exception.custom.ExistedDuplicateValueException;
 import roomescape.exception.custom.InUseException;
 import roomescape.exception.custom.NotExistedValueException;
